+++ src/store/uninstallUtils.ts
